add insert and getSize methods

+++ src/main/kotlin/pro/komdosh/kot/multiqueue/api/Multiqueue.kt
@@ -3,7 +3,7 @@ package pro.komdosh.kot.multiqueue.api
 interface Multiqueue<T> {
     fun printSize()
 
-    fun getSize(): Int
+    fun getSize(): Long
 
     suspend fun balance()
